+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Minor
 
 * Internal: Turn on sketchy-number flow lint rules as an error (#293)
+* TextArea: Add an onKeyDown prop (#303)
+* TextField: Add an onKeyDown prop (#303)
 * Internal: Add flow types to `Box` transform functors (#299)
 * Icon: Fix cog icon rotation (#308)